requirements.txt
react-router-dom@6.21.1
react@18.2.0
tailwindcss@3.3.6
Flask@2.3.3
python-dotenv@0.19.2
psycopg2-binary
flask-cors